Standalone grid implemented.
