+++ lib/index.ts
@@ -2,8 +2,6 @@
 export * from './tubular.module';
 
 // export services
-export * from './tubular-auth.service';
-export * from './tubular-http.service';
 export * from './tubular-settings.service';
 
 // export components and directives
