+++ sample/app/app.component.html
@@ -9,9 +9,6 @@
             <a class="nav-link" routerLink="/grid" routerLinkActive="active">Grid Sample</a>
         </li>
         <li class="nav-item">
-            <a class="nav-link" routerLink="/form" routerLinkActive="active">Form Sample</a>
-        </li>
-        <li class="nav-item">
             <a class="nav-link" routerLink="/login" routerLinkActive="active">Login</a>
         </li>
         <li class="nav-item">
