+++ sample/app/app.module.ts
@@ -3,7 +3,7 @@ import { BrowserModule } from '@angular/platform-browser';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { RequestOptions } from '@angular/http';
 
-import { TubularModule, SETTINGS_PROVIDER, TubularLocalStorageService, HttpOptions, TubularAuthService, TubularHttpService } from '@tubular2/tubular2';
+import { TubularModule, SETTINGS_PROVIDER, TubularLocalStorageService, HttpOptions } from '@tubular2/tubular2';
 
 import { AppRoutingModule } from './app-routing.module';
 
@@ -16,15 +16,13 @@ import { ToastModule } from 'ng2-toastr/ng2-toastr';
 @NgModule({
     imports: [
         BrowserModule, FormsModule, ReactiveFormsModule, TubularModule, 
-        AppRoutingModule, ToastModule
+        AppRoutingModule, ToastModule.forRoot()
     ],
     declarations: [
         AppComponent, SampleGrid, MainGridComponent, 
         LoginComponent
     ],
     providers: [
-        TubularAuthService,
-        TubularHttpService,
         { provide: SETTINGS_PROVIDER, useClass: TubularLocalStorageService },
         { provide: RequestOptions, useClass: HttpOptions }
     ],
