+++ sample/app/login.component.ts
@@ -1,9 +1,12 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, Injectable, Inject, Optional } from '@angular/core';
+import { Http, RequestOptions, RequestOptionsArgs } from '@angular/http';
 import { Router } from '@angular/router'
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 
-import { TubularAuthService } from '@tubular2/tubular2';
+import { SETTINGS_PROVIDER, ITubularSettingsProvider } from '@tubular2/tubular2';
 import { ToastsManager } from 'ng2-toastr/ng2-toastr';
+
+import 'rxjs/add/operator/map';
 
 @Component({
     selector: 'login',
@@ -13,7 +16,22 @@ import { ToastsManager } from 'ng2-toastr/ng2-toastr';
 export class LoginComponent implements OnInit {
     loginForm: FormGroup;
 
-    constructor(private fb: FormBuilder, private tbAuthService: TubularAuthService, private router: Router, private toastr: ToastsManager) { }
+    private userData = {
+        isAuthenticated: false,
+        username: '',
+        bearerToken: '',
+        expirationDate: null,
+        role: '',
+        refreshToken: ''
+    };
+
+    constructor(
+        @Optional() @Inject(SETTINGS_PROVIDER) private settingsProvider: ITubularSettingsProvider,
+        private fb: FormBuilder,
+        private http: Http,
+        private router: Router,
+        private toastr: ToastsManager) {
+    }
 
     ngOnInit() {
         this.loginForm = this.fb.group({
@@ -26,16 +44,39 @@ export class LoginComponent implements OnInit {
         let username = data.value.username;
         let password = data.value.password;
 
-        this.tbAuthService.authenticate(username, password)
+        let requestArgs = <RequestOptionsArgs>{
+            headers: new Headers(
+                {
+                    'Content-Type': 'application/x-www-form-urlencoded'
+                })
+        }
+
+        this.http.post('http://tubular.azurewebsites.net/api/token', `grant_type=password&username=${username}&password=${password}`, requestArgs)
+            .map((data) => this.handleSuccessCallback(data))
             .subscribe(
             data => {
                 console.log("Authenticated", data);
                 this.router.navigate(['/exp']);
             },
             error => {
-                this.toastr.error(error.status + ' - ' + error.errorBody.error_description, 'Filed Login');
+                let errorJson = error.json();
+                alert(error.status + ' - ' + errorJson.error_description);
                 this.router.navigate(['/login']);
             }
             );
+    }
+
+    private handleSuccessCallback(data) {
+        data = JSON.parse(data._body);
+        this.userData.isAuthenticated = true;
+        this.userData.username = data.userName;
+        this.userData.bearerToken = data.access_token;
+        this.userData.expirationDate = new Date(new Date().getTime() + data.expires_in * 1000);
+        this.userData.role = data.role;
+        this.userData.refreshToken = data.refresh_token;
+
+        if (this.settingsProvider) {
+            this.settingsProvider.put('auth_data', JSON.stringify(this.userData));
+        }
     }
 }
