+++ sample/app/main-grid.component.html
@@ -4,10 +4,7 @@
     Angular2 components
 </p>
 
-<tb-grid #grid 
-    dataUrl="http://tubular.azurewebsites.net/api/orders/paged" 
-    requireAuthentication="false" 
-    (onDataError)="errorHandler($event)">
+<tb-grid #grid dataUrl="http://tubular.azurewebsites.net/api/orders/paged" requireAuthentication="false" (onDataError)="errorHandler($event)">
     <div class="row">
         <div class="col-md-5">
             <tb-grid-pager></tb-grid-pager>
@@ -16,9 +13,9 @@
             <div class="btn-group-sm">
                 <tb-grid-print></tb-grid-print>
                 <tb-grid-export fileName="sampleCsv.csv"></tb-grid-export>
-                <button class="btn btn-info" (click)="openDialog(content)">
+                <!--<button class="btn btn-info" (click)="openDialog(content)">
                     New Item
-                </button>
+                </button>-->
             </div>
         </div>
         <div class="col-md-3">
