+++ sample/app/sampleGrid.component.html
@@ -19,12 +19,12 @@
     <tbody>
         <tr *ngFor="let row of rows">
             <td>
-                <button class="btn btn-primary btn-sm" (click)="details(row)">
+                <!--<button class="btn btn-primary btn-sm" (click)="details(row)">
                     Details
                 </button>
                 <button class="btn btn-info btn-sm" (click)="editRow(content)">
                     Edit 
-                </button>
+                </button>-->
             </td>
             <td>{{row.OrderID}}</td>
             <td>{{row.CustomerName}}</td>
